### 2016-09-12 #61 @whedon
[REVIEW]: walkr: MCMC Sampling from Non-Negative Convex Polytopes
**Submitting author:** @andyyao95 (<a href="http://orcid.org/0000-0003-3898-8782">Andy  Yao</a>)
**Repository:** <a href="https://github.com/andyyao95/walkr" target ="_blank">https://github.com/andyyao95/walkr</a>
**Version:** v0.3.4

**Reviewers and authors**: Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue
in the <a href="https://github.com/andyyao95/walkr" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers)
in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice
versa will create corresponding breadcrumb trails in the link target.)
