### 2016-09-12 #61 @arfon
Hi @andyyao95. Currently the paper doesn't compile properly as you're not actually citing any of the references in the paper. Could you actually cite the entries in your `paper.bib` (e.g. [@kannan]) in the body of the paper?
