+++ paper.bib
@@ -1,113 +1,0 @@
-@article{kannan,
-         author = {Kannan, Ravi and Narayanan, Hariharan},
-         title = {{Random Walks on Polytopes and an Affine Interior Point Method for Linear Programming}},
-         journal = {Mathematics of Operations Research},
-         year = {2009}
-}
-
-@manual{hitandrun, 
-        author = {van Valkenhoef, Gert and Tervonen, Tommi},
-        title = {hitandrun: "Hit and Run" and "Shake and Bake" for Sampling Uniformly from Convex Shapes},
-        organization = {CRAN},
-        year = {2015}
-}
-
-@Article{rcpp,
-    title = {{Rcpp}: Seamless {R} and {C++} Integration},
-    author = {Dirk Eddelbuettel and Romain Fran\c{c}ois},
-    journal = {Journal of Statistical Software},
-    year = {2011},
-    volume = {40},
-    number = {8},
-    pages = {1--18},
-    url = {http://www.jstatsoft.org/v40/i08/},
-  }
-
-@article{rcppeigen,
-    title = {Fast and Elegant Numerical Linear Algebra Using the
-      {RcppEigen} Package},
-    author = {Douglas Bates and Dirk Eddelbuettel},
-    journal = {Journal of Statistical Software},
-    year = {2013},
-    volume = {52},
-    number = {5},
-    pages = {1--24},
-    url = {http://www.jstatsoft.org/v52/i05/}
-}
-
-@article{gelman,
-    title = {Inference from Iterative Simulation
-    Using Multiple Sequences},
-    author = {Andrew Gelman and Donald B. Rubin},
-    journal = {Statistical Science},
-    year = {1992},
-    volume = {7},
-    number = {4},
-    pages = {457--511},
-    url = {https://projecteuclid.org/download/pdf_1/euclid.ss/1177011136}
-}
-
-@book{leon,
-    title = {Linear Algebra with Applications},
-    author = {Steven J. Leon},
-    publisher = {Pearson}, 
-    year = {2014}
-}
-
-@article{vempala,
-  title = {Geometric Random Walks: A Survey}, 
-  author = {Santosh Vempala}, 
-  journal = {Combinatorial and Computational Geometry},
-  year = {2005}, 
-  volume = {52}, 
-  pages = {573--612}
-}
-
-@Article{ravi, 
-         author = {Kannan, Ravi and Lovasz, Laszlo},
-         title = {{Random Walks and a Volume Algorithm for Convex Bodies}},
-         journal = {Random Structures and Algorithms},
-         year = {1997}
-}
-@Article{limsolve,
-    author = {Karel Van den Meersche and Karline Soetaert and Dick Van
-      Oevelen},
-    title = {xsample(): An R Function for Sampling Linear Inverse
-      Problems},
-    journal = {Journal of Statistical Software, Code Snippets},
-    volume = {30},
-    number = {1},
-    pages = {1--15},
-    year = {2009},
-    coden = {JSSOBK},
-    issn = {1548-7660},
-    url = {http://www.jstatsoft.org/v30/c01},
-}
-
-@Article{gelmanrubin,
-  author = {Andrew Gelman and Donald Rubin}, 
-  title = {Inference from Iterative Simulation Using Multiple Sequences}, 
-  journal = {Statistical Science}, 
-  volume = {7}, 
-  number = {4}, 
-  pages = {457--511},
-  year = {1992},
-}
-
-@manual{stanmanual, 
-        author = {{Stan Development Team}},
-        title = {Stan Modeling Language: User’s Guide and Reference Manual},
-        organization = {Stan},
-        year = {2015}
-}
-
-@manual{shinystan,
-        author = {Jonah Gabry},
[…]
