### 2016-09-12 #61 @andyyao95
Hi Arfon,

Thank you so much, I used the reference for a longer version of the paper
instead (my bad). I just updated the paper.bib and paper.md to properly
cite all the references.

Let me know if you have any other questions. Thanks!

Andy

On Mon, Sep 12, 2016 at 12:54 AM, Arfon Smith <notifications@github.com>
wrote:

> Hi @andyyao95 <https://github.com/andyyao95>. Currently the paper doesn't
> compile properly as you're not actually citing any of the references in the
> paper. Could you actually cite the entries in your paper.bib (e.g. [
> @kannan <https://github.com/kannan>]) in the body of the paper?
>
