### 2016-09-21 #61 @jakevdp
I contacted Neal Thomas, author of R2OpenBUGS, to see if he's interested in reviewing.
