### 2016-09-23 #61 @jakevdp
Neal was not interested, nor were a couple other R/Bayesian folks. Still reaching out to some...
