### 2016-09-23 #61 @karthik
@jakevdp You might ask @drgtwo if he's interested and has the time.
