### 2016-09-29 #61 @jakevdp
Hi folks,
I'm completely striking out finding a reviewer for this. To be entirely honest, I accepted this thinking it was a Python package. I'm finding I don't have enough ties to the Rstats community to find an appropriate person/someone who knows me well enough to not blow me off.

Is there a way to re-assign this to another editor? I'm frankly out of ideas.
