### 2017-01-08 #61 @arfon
> @dgrtwo Hi Dave, can you update this thread when you get a chance? Thanks!

Friendly new year reminder @dgrtwo 😁
