### 2017-02-05 #61 @arfon
:wave @dgrtwo - are you still able to review this submission for JOSS? If not, we can look for another reviewer.
