### 2017-02-16 #61 @masalmon
@arfon just noticed this from the CRANberries feed https://cran.r-project.org/web/packages/walkr/index.html
