### 2017-02-16 #61 @masalmon
The reason for removal from CRAN is "Archived on 2017-02-16 as check problems were not corrected despite reminders."
