### 2017-02-17 #61 @andyyao95
@arfon @masalmon  Hi. Very sorry about this. There was a minor error in compiling the package vignette on CRAN. I missed the reminder as it was in another email.

I've updated the package (same content, small vignette fix). I'll let you guys know when CRAN updates it. Thanks!
