### 2017-02-17 #61 @masalmon
@andyyao95 great I had just copied this here in case you were MIA :wink: cool that you could solve the error!
