### 2017-02-24 b5550ee
update DESCRIPTION per CRAN documentation
