+++ DESCRIPTION
@@ -1,29 +1,0 @@
-Package: walkr
-Version: 0.3.3.9001
-Date:    2015-08-27
-Type:    Package
-Title:   Random Walks in the Intersection of Hyperplanes and the N-Simplex 
-Author:  as.person(c(
-                "Andy Yao <andy.yao17@gmail.com> [aut]",
-                "David Kane <dave.kane@gmail.com> [aut]"                
-              ))
-Maintainer:  Andy Yao <andy.yao17@gmail.com>
-Depends:     R (>= 3.1), ggplot2
-Suggests:    testthat, knitr, grid, gridExtra
-Imports:     hitandrun, 
-             limSolve, 
-             MASS,
-             Rcpp (>= 0.11.6),
-             shinystan,
-             parallel
-Description: Consider the intersection of two spaces: the complete solution space
-             to Ax = b and the N-simplex. The intersection of these two spaces is 
-             a non-negative convex polytope. The package walkr samples from this 
-             intersection using two Monte-Carlo Markov Chain (MCMC) methods: 
-             hit-and-run and Dikin walk. walkr also provide tools to examine sample 
-             quality.
-License:     GPL-3
-URL:         https://github.com/andyyao95/walkr
-BugReports:  https://github.com/andyyao95/walkr/issues
-LinkingTo:   Rcpp, RcppEigen
-VignetteBuilder: knitr
