### 2017-02-24 5326925
update DESCRIPTION per CRAN documentation
