### 2017-02-24 #61 @andyyao95
It's back on CRAN [now.](https://cran.r-project.org/web/packages/walkr/index.html) (also updated on github.) Thank you guys for your patience. Let me know if anything else is needed in the review process.

Looking forward to hearing back, and have a nice day!
