### 2017-03-16 75fc1a3
update to latest CRAN version
