### 2017-03-16 1ade63e
update to latest CRAN version
