### 2017-03-16 b16eb30
update to latest CRAN version
