### 2017-03-16 abb7faf
update to latest CRAN version
