+++ R/kmatch.R
@@ -1,110 +1,0 @@
-#' Create matching vectors based on data 
-#' 
-#' Creates a matrix of weights which match a given set of weights in exposures
-#' to designated factors
-#' 
-#' @param x data frame containing needed input data
-#' @param weight.var character name of the column of the input weights
-#' @param match.var character vector of names of columns of 'data' we wish to match on
-#' @param n numeric number of weight vectors desired. Default is 1
-#' @param thin only 1 out of every 'thin' vectors will be recorded. Default is 5
-#' @param replace if FALSE then any rows that had non-zero weight in 'weight.var' will
-#' be set to 0 in the output. Default is FALSE
-#' @param check whether to remove linearly dependent rows from A. Default is TRUE
-#' @param chains number of chains. Default is 1.
-#' 
-#' @return Returns a list of "chains" matrices of 'n' sets of weights that match the given set of
-#' weights in terms of weighted averages to the 'match.var' factors. The columns of each matrix
-#' are the sets of weights.
-#' @export
-#' @author Mike Flynn \email{mflynn210@@gmail.com}, Yuanchu Dang \email{yuanchu.dang@gmail.com}
-#'   
-#'   
-#' @examples
-#' set.seed(40)
-#' x <- data.frame(size = rnorm(50), weight = rep(.02, 50))
-#' weights <- kmatch(x, weight.var = "weight", match.var = "size", n = 100, replace = TRUE)
-#' 
-#' x2 <- data.frame(var1 = c(-2,-1, 0, 1, 2), var2 = c(1, -1, 2, -1, 0), w = rep(.2, 5))
-#' matchvars = c("var1", "var2")
-#' weights2 = kmatch(x = x2, weight.var = "w", match.var = matchvars, n = 100, thin = 100, replace= TRUE)
-#' ## for interactive graphics:
-#' ## library(rgl)
-#' ## plot3d(x = weights2[1,], y = weights2[2,], z = weights2[3,])
-
-kmatch <- function(x,
-                   weight.var,
-                   match.var,
-                   n = 1,
-                   thin = 5,
-                   replace = FALSE,
-                   check = TRUE,
-                   chains = 1) {
-  
-  ## Input checking.
-  
-  stopifnot(is.data.frame(x))
-  stopifnot(is.character(weight.var))
-  stopifnot(is.character(match.var))
-  stopifnot(n > 0)
-  stopifnot(all(c(weight.var, match.var) %in% names(x)))
-  
-  notincluded = numeric()
-  ret = matrix(0, nrow = nrow(x), ncol = n)
-  
-  if(any(is.na(x[[weight.var]]))) {
-    warning(paste(length(which(is.na(x[[weight.var]]))), " entries of '", weight.var, "' are missing, these rows are zeroed in the output", sep = ""))
-    notincluded = c(notincluded, which(is.na(x[[weight.var]])))
-    x = subset(x, !is.na(x[[weight.var]]))
-  }
-  
-  for(v in match.var) {
-    if(any(is.na(x[[v]]))) {
-      warning(paste(length(which(is.na(x[[v]]))), " entries of '", v, "' are missing, these rows are zeroed in the output", sep = ""))
-      notincluded = c(notincluded, which(is.na(x[[v]])))
-      x = subset(x, !is.na(x[[v]]))
-    }
-  }
-  
-  ## Might consider moving the bulk of these inline comments to the description
-  ## above, with better math formatting. For now, I will just leave them here.
-  
-  ## First step is to build the constraint equation that samples will need to 
-  ## satisfy: Ax = b. Note that this equation does not include the inequality 
-  ## constraint that all elements of x must be >= 0. A is a matrix which 
-  ## includes the match.var variable exposures as rows. Multiplying these rows 
[…]
