### 2017-03-16 a6a6a36
update to latest CRAN version
