### 2017-03-16 275afec
update to latest CRAN version
