### 2017-03-16 a944956
update to latest CRAN version
