### 2017-03-16 b021e9e
update R files to latest CRAN version
