+++ R/dikin_walk.R
@@ -32,9 +32,8 @@ dikin_walk <- function(A,
                        r = 1,
                        thin = 1,
                        burn = 0,
-                       chains = 1,
-                       c) {
-
+                       chains = 1) {
+  
   stopifnot(points %% chains == 0)
   stopifnot(is.list(x0))
   #############################
@@ -52,26 +51,15 @@ dikin_walk <- function(A,
   ## H(x) is the Hessian of the Log-barrier function of Ax <= b
   ## for more details, just google Dikin Walk
   
-  H_extra <- function (c, x) {
-    
-    x <- as.vector(x)
-    
-    # function returns Hessian of exp(X-C)
-    distance <- sum((x - c)^2)
-    return(4*exp(distance)*(x - c) %*% t(x - c) +
-             2*exp(distance)*diag(length(x)))  
-  }
-  
-  H_x <- function(x, c) {
+  H_x <- function(x) {
     
     ## making the D^2 matrix
-    
+
     D <- as.vector(1/(A_b[,1] - rcppeigen_fprod(A_b[,-1], x)))
-    
+
     ## t(A) %*% (D^2 %*% A)
     
-    return(rcppeigen_fcrossprod(A, rcppeigen_fprod(diag(D^2), A)) + 
-             H_extra(c, x))
+    return(rcppeigen_fcrossprod(A, rcppeigen_fprod(diag(D^2), A)))
     
   } 
   
@@ -112,32 +100,128 @@ dikin_walk <- function(A,
 #     2) should move these 3 functions into separate files with documentation...
 #     
     
-    return(as.numeric(rcppeigen_fcrossprod(z-x, rcppeigen_fprod(H_x(x, c), (z-x)))) <= r^2)
+    return( as.numeric(rcppeigen_fcrossprod(z-x, rcppeigen_fprod(H_x(x), (z-x)))) <= r^2)
     
   } 
   
   ###### THE LINES ABOVE FINISH DEFINING THE ELLIPSOID
-  ## total points is : points * thin * 1/(1-burn) / chains 
-  ## because burn-in is a percentage, we must take the CEILING function
-  ## to sample more than we need (in the case where dividing by 1-burn
-  ## does not return an integer)
+  ## now the sampling
   
-  total.points <- ceiling( (points / chains)  * thin * (1/(1-burn))) 
+  ## initialize return matrix
+  ## set the starting point as the current point
+  answer <- list()
   
-  # check how many cores the machine has
-  numCores <- detectCores()
-  
-  # use one less number of cores to be sensible
-  cl <- makeCluster(numCores - 1)
+  ##total points for each indiv chain
   
   for (j in 1:chains) {
-    # parallel by using parLapply
-    answer <- parLapply(cl, x0, Dikin_single_chain, total.points,
-                        A, b, r, A_b, burn, chains, points, thin, c) 
-  }
   
-  # stop the cluster after the parallel
-  stopCluster(cl) 
+    ## total points is : points * thin * 1/(1-burn) / chains 
+    ## because burn-in is a percentage, we must take the CEILING function
+    ## to sample more than we need (in the case where dividing by 1-burn
+    ## does not return an integer)
+    
+    total.points <- ceiling( (points / chains)  * thin * (1/(1-burn))) 
+    
+    ## initializing the return matrix 
+    
+    result <- matrix(ncol = total.points, nrow = ncol(A))
+    result[ , 1] <- x0[[j]]
+    current.point <- x0[[j]]
+    this.length <- length(b)
+    
+    for (i in 2:total.points) {
+      
+      ## 1. Generate random point y in Ellip(x)
+      ## KEY: MUST USE STANDARD NORMAL FUNCTION HERE, RUNIF IS NOT UNIFORM AFTER TRANSFORMATION
+      ## see vignette for details
+      
[…]
