### 2017-03-16 0bb4394
update to CRAN, march 2017
