+++ man/rcppeigen_fprod.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/RcppExports.R
 \name{rcppeigen_fprod}
 \alias{rcppeigen_fprod}
@@ -21,5 +21,5 @@ Computes using RcppEigen the product of A and B
 \dontrun{
 rcppeigen_fprod(A, B)
 }
-}
 
+}
