+++ man/hit_and_run.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/hit_and_run.R
 \name{hit_and_run}
 \alias{hit_and_run}
@@ -26,15 +26,14 @@ a list of chains of the sampled points, each chain
         being a matrix object with each column as a point
 }
 \description{
-This function provides a wrapper
-for the har function of the
+This function provides a wrapper 
+for the har function of the 
 hit-and-run package
 }
 \examples{
 \dontrun{
-## note that this Ax <= b is different from Ax=b that the
-## user specifies for walkr (see transformation section in vignette)
-hit_and_run(A = A, b = b, x0, points = 100, thin = 1, burn = 0, chains = 1)
+## note that this Ax <= b is different from Ax=b that the 
+## user specifies for walkr (see transformation section in vignette) 
+hit_and_run(A = A, b = b, x0, points = 100, thin = 1, burn = 0, chains = 1) 
+} 
 }
-}
-
