+++ man/explore_walkr.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/explore_walkr.R
 \name{explore_walkr}
 \alias{explore_walkr}
@@ -8,8 +8,8 @@ explore_walkr(x)
 }
 \arguments{
 \item{x}{is the set of points sampled, with its columns as the sampled points.
-       If multiple chains are present, then the columns should be ordered
-       such that each chain follow each other.}
+If multiple chains are present, then the columns should be ordered
+such that each chain follow each other.}
 }
 \value{
 a shiny interface that display the diagnostics of the MCMC random walk
@@ -21,4 +21,3 @@ interval of each dimension's coordiates across the whole sets of points,
 Gelman-Rubin statistics, trace-plots, and other diagnostic tools for
 examining convergence.
 }
-

[…]
