### 2017-03-16 351b386
update bib for DOI
