+++ paper.bib
@@ -1,0 +1,43 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% http://bibdesk.sourceforge.net/
+
+
+%% Created for Andy at 2016-09-12 16:34:03 -0400 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
+@article{kannan,
+	Author = {Kannan, Ravi and Narayanan, Hariharan},
+	Journal = {Mathematics of Operations Research},
+	Title = {{Random Walks on Polytopes and an Affine Interior Point Method for Linear Programming}},
+	doi = {10.1287/moor.1110.0519},
+	Year = {2009}}
+
+@article{gelman,
+	Author = {Andrew Gelman and Donald B. Rubin},
+	Journal = {Statistical Science},
+	Number = {4},
+	Pages = {457--511},
+	Title = {Inference from Iterative Simulation Using Multiple Sequences},
+	Url = {https://projecteuclid.org/download/pdf_1/euclid.ss/1177011136},
+	doi = {10.1214/ss/1177011136}
+	Volume = {7},
+	Year = {1992},
+	Bdsk-Url-1 = {https://projecteuclid.org/download/pdf_1/euclid.ss/1177011136}}
+
+@article{vempala,
+	Author = {Santosh Vempala},
+	Journal = {Combinatorial and Computational Geometry},
+	Pages = {573--612},
+	Title = {Geometric Random Walks: A Survey},
+	doi = {10.1.1.170.9791},
+	Volume = {52},
+	Year = {2005}}
+
+@manual{shinystan,
+	Author = {Jonah Gabry},
+	Title = {{Interactive Visual and Numerical Diagnostics and Posterior Analysis for for Bayesian Models}},
+	Year = {2015}}
