### 2017-03-16 andyyao95/walkr#2 @andyyao95
Hi @dgrtwo,

Thank you for your time and effort in reviewing our paper! With regard to your comments:

- I just updated the github version and it is now the same version as the one on CRAN (I should've done this earlier). The github version now should be able to pass all tests and CMD checks, as it did on my machine. Let me know if there is any issue with this.

- I've updated the .bib file, including all the DOIs I could find, except one which is a reference to an R package.

Thanks again. Let me know if you have any other questions. Looking forward to hearing back from you.

Best,
Andy
