### 2017-03-16 andyyao95/walkr#2 @dgrtwo
Great! I hope you do complete the other vignettes when you have the opportunity, they seem valuable.

Can you submit a new GitHub release for 0.3.4 tagged to this commit? At that point I'll recommend acceptance.
