### 2017-03-16 andyyao95/walkr#2 @andyyao95
Hi @dgrtwo,

Just submitted a release for 0.3.4! Thanks.

Andy
