### 2017-03-16 #61 @dgrtwo
After review I'm recommending **acceptance**; my only issues (about GitHub/CRAN differences and DOIs) were resolved in [this issue](https://github.com/andyyao95/walkr/issues/2). My other note is that the version is now **0.3.4** (I don't know where that should change in the submission).

Overall this is a great package and I've been impressed as I've played with the mathematical sampling. I particularly like the detail and figures in the vignette and am impressed with the test coverage. paper.md covers the software's uses well.

I greatly appreciate your patience in this unusually long process, as well as Andy's quick response to feedback!
