### 2017-03-20 #61 @arfon
@davidkane9 - https://github.com/openjournals/joss-papers/raw/master/joss.00061/10.21105.joss.00061.pdf _is_ the full JOSS paper. Please take a look at the author guidelines [here](http://joss.theoj.org/about#author_guidelines).
