### 2017-03-20 #61 @davidkane9
Ah, yes. Read that so long ago that I had forgotten it!

;-)

On Mon, Mar 20, 2017 at 9:23 AM, Arfon Smith <notifications@github.com>
wrote:

> @davidkane9 <https://github.com/davidkane9> - https://github.com/
> openjournals/joss-papers/raw/master/joss.00061/10.21105.joss.00061.pdf
> *is* the full JOSS paper. Please take a look at the author guidelines here
> <http://joss.theoj.org/about#author_guidelines>.
>
